refactor: clean up OTP constants, deduplicate test helpers, parallelize route queries
- rename OTP_LENGTH and CHALLENGE_TOKEN_HEADER to camelCase per project conventions
- extract shared createSpyHttpClient and jsonResponse into test-utils/http-client.ts
- parallelize challenge token creation and user lookup in waitlist route
- replace magic number 8 with otpLength constant in waitlist page
- remove empty table constraint callback from OTP schema
- add console.info to previously empty catch block
