+++ src/components/sign-in/use-sign-in-form-state.ts
@@ -1,5 +1,5 @@
 import type { AuthClient } from '@/contexts'
-import { CHALLENGE_TOKEN_HEADER, OTP_LENGTH } from '@/lib/constants'
+import { challengeTokenHeader, otpLength } from '@/lib/constants'
 import { HttpError, type HttpClient } from '@/lib/http'
 import { getOtpErrorMessage } from '@/lib/otp-error-messages'
 import { updateSettings } from '@/dal'
@@ -17,8 +17,8 @@ const getServerErrorMessage = async (error: unknown, fallback: string): Promise<
     if (typeof body?.message === 'string' && body.message) {
       return body.message
     }
-  } catch {
-    // Response body not JSON-parseable
+  } catch (e) {
+    console.info('Could not parse error response body as JSON:', e)
   }
   return fallback
 }
@@ -180,7 +180,7 @@ export const useSignInFormState = ({
   }
 
   const handleOtpComplete = async (value: string) => {
-    if (value.length !== OTP_LENGTH) {
+    if (value.length !== otpLength) {
       return
     }
 
@@ -191,7 +191,7 @@ export const useSignInFormState = ({
         email: state.email.trim(),
         otp: value,
         fetchOptions: {
-          headers: { [CHALLENGE_TOKEN_HEADER]: state.challengeToken },
+          headers: { [challengeTokenHeader]: state.challengeToken },
         },
       })
